+++ classes/Plugin/Core.php
@@ -292,6 +292,18 @@ final class Core extends \PopupMaker\Plugin\Container {
 			}
 		);
 
+		$this->set(
+			'link_click_tracking',
+			/**
+			 * Get link click tracking service.
+			 *
+			 * @return \PopupMaker\Services\LinkClickTracking
+			 */
+			function ( $container ) {
+				return new \PopupMaker\Services\LinkClickTracking( $container );
+			}
+		);
+
 		do_action( 'popup_maker/register_services', $this );
 	}
 
@@ -302,6 +314,10 @@ final class Core extends \PopupMaker\Plugin\Container {
 	 */
 	protected function init_services() {
 		$license = $this->get( 'license' );
+
+		// Initialize link click tracking.
+		$link_click_tracking = $this->get( 'link_click_tracking' );
+		$link_click_tracking->init();
 	}
 
 	/**